% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shinyapp.R
\name{Eduexp}
\alias{Eduexp}
\title{Shiny app ue an interactive interface to show the expenditure in education in Italy and its closest countries.}
\usage{
Eduexp()
}
\description{
The function goes without arguments
}
