% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Appshiny.R
\name{ExpEdu}
\alias{ExpEdu}
\title{ExpenditureEductation}
\usage{
ExpEdu()
}
\value{
shiny app
}
\description{
This function permits to the user to launch the shiny app.
Through this app, there will be printed out three different graphs, depending on the choice of the user. User, in fact, can choose three different years and according to them, shiny will run a different plot for the countries in consideration.
}
