% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getdata.R
\name{getData}
\alias{getData}
\title{Acquisition of data}
\usage{
getData()
}
\description{
this function permits the users to get the data from a csv file stored in the folder "inst/extdata".
Putting this function into the shiny app, it will get the data needed for the package.
}
