% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Education expenditure data.R
\name{importData}
\alias{importData}
\title{Import data from the World bank database.}
\usage{
importData()
}
\description{
This function returns the data of the Expenditure in education of all countries all over the world.
Data source : https://data.worldbank.org/indicator/SE.XPD.TOTL.GD.ZS
}
